+++ console-socket.c
@@ -34,7 +34,8 @@ ssize_t console_socket_path(socket_path_t sun_path, const char *id)
 	ssize_t rc;
 
 	if (!id) {
-		id = "default";
+		errno = EINVAL;
+		return -1;
 	}
 
 	rc = snprintf(sun_path + 1, sizeof(socket_path_t) - 1,